Fixes #4295 - Removing provider subcommands
This is because we are moving to a centralized anonymous/redhat
provider. Every org is going to have one of each of these and we are not
going to let you perform crud on those. Also all the commands provided
by provider can be simulated via product list or subscriptions import.