Fixes #9819 - No default values for Sync Plans.
If you pass ``none`` as the interval for a Sync Plan you get an error.
Turns out not even the API or UI allow you to create a sync plan with no
interval set.

I have updated the code to no longer list ``none`` as a valid option and
set ``weekly`` as the new default value.

Also, a default value is no longer provided for the ``sync_date`` field.