fixes #11968: add content name to repository information returned
Currently, when a user invokes the 'repository-set available-repositories'
command, the 'name' field is displayed simmilar to the following:
  Red Hat Enterprise Linux 6 Server RPMs x86_64 6Server

The above includes the architecture and release version; however, that
information is also returned as separate attributes.  Since the name
can be used for operations such as enable, disable, info, it would
be better to return the name similar to the following:
  Red Hat Enterprise Linux 6 Server (RPMs)

This commit displays the 'name' vs the 'repo_name' attribute returned.