Fixes #14853 - Support organization options in host-collections
As usual, we require organization only when a host collection name is
specified rather than an ID. This is not true of `host-collection list`
command, as it will list all host collections across all organizations
if no organization options are given.

In addition, the host-collection copy interface is updated to be
consistent with the rest of hammer, allowing searching for a host
collection by name if the organization is also specified. The new host
collection name is now specified with --new-name rather than --name.

Note that I added test files for the broken content-host subcommand and
created an issue for fixing the broken commands
[#15289](http://projects.theforeman.org/issues/15289).

There is also another broken command `host-collection hosts`. The bug
report to fix that is
[#15290](http://projects.theforeman.org/issues/15290).

The host-collection add-host/remove-host commands are broken as well,
always reporting success even when the hosts in question were not
found: [#15291](http://projects.theforeman.org/issues/15291) addresses
this bug.