fixes #21371 - convert repository-set enable/disable to be Command (#520)
The repository-set enable/disable commands are currently
derived from the UpdateCommand.  This is resulting in the
inclusion of a '--new-name' parameter, which is not needed
or used by the commands.  As a result, updating them
to be simple Commands so that the parameter is not included.

(cherry picked from commit 048c526ac2ae084e46fd28e5829629278885b428)