Fixes #25272 - add support for composite CV import/export
Previously, CV import/export would treat composites as "regular"
content views.  This resulted in each repo of the composite being
exported as if it was part of a regular content view.

These exports were unable to be imported since the repos existed on
the components and not the composite.

This commit exports composite content views with information needed to
reconstitute them on the importing side. Each component CV name and
version is written to the JSON, and upon import, hammer searches for
the matching name and version on the importing side. If all of the
info lines up, the components are associated with the CV and a publish
occurs.