Fixes #25775 - create content view if it doesn't exist
Previously, the content view had to be created ahead of time before
content view versions were imported.

This patch creates the content view if it doesn't already exist. Note
that it only creates the CV with name and repos, it does not do
activation keys. Composites are handled as well.